requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
cachecontrol==0.14.2; python_version >= '3.8'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.1.31; python_version >= '3.6'
cffi==1.17.1; platform_python_implementation != 'PyPy'
channels==4.2.0; python_version >= '3.8'
channels-redis==4.2.1; python_version >= '3.8'
charset-normalizer==3.4.1; python_version >= '3.7'
cryptography==44.0.1
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-rest-auth[with_social]==7.0.1; python_version >= '3.8'
django==5.1.6; python_version >= '3.10'
django-allauth[socialaccount]==65.4.1
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt[crypto]==5.4.0; python_version >= '3.9'
djoser==2.3.1; python_version >= '3.8' and python_version < '4.0'
firebase-admin==6.6.0; python_version >= '3.7'
google-api-core[grpc]==2.24.1; platform_python_implementation != 'PyPy'
google-api-python-client==2.162.0; python_version >= '3.7'
google-auth==2.38.0; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2; python_version >= '3.7'
google-cloud-firestore==2.20.0; platform_python_implementation != 'PyPy'
google-cloud-storage==3.0.0; python_version >= '3.7'
google-crc32c==1.6.0; python_version >= '3.9'
google-resumable-media==2.7.2; python_version >= '3.7'
googleapis-common-protos==1.68.0; python_version >= '3.7'
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0; python_version >= '3.7'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.10; python_version >= '3.6'
msgpack==1.1.0; python_version >= '3.8'
oauthlib==3.2.2; python_version >= '3.6'
packaging==24.2; python_version >= '3.8'
proto-plus==1.26.0; python_version >= '3.7'
protobuf==5.29.3; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.1; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyjwt[crypto]==2.10.1; python_version >= '3.9'
pyparsing==3.2.1; python_version >= '3.1'
python3-openid==3.2.0
redis==5.2.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0
rsa==4.9; python_version >= '3.6' and python_version < '4'
social-auth-app-django==5.4.3; python_version >= '3.9'
social-auth-core==4.5.6; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
tzdata==2025.1; sys_platform == 'win32'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.3.0; python_version >= '3.9'
